move overlay to grid wrapper
