+++ frontend/src/components/GridWrapper/styles.scss
@@ -1,0 +1,18 @@
+.grid-wrapper {
+    &__grid {
+        position: relative; // For the overlay
+    }
+
+    &__overlay {
+        position: absolute;
+        top: 0;
+        left: 0;
+        bottom: 0;
+        right: 0;
+        z-index: -1;
+
+        &_visible {
+            z-index: 1;
+        }
+    }
+}
